Fix macOS build.

+++ src/b2/DiscGeometry.h
@@ -57,12 +57,12 @@ extern const std::vector<std::string> DISC_IMAGE_EXTENSIONS;
 
 bool FindDiscGeometryFromFileDetails(DiscGeometry *geometry,
                                      const char *file_name,
-                                     size_t file_size,
+                                     uint64_t file_size,
                                      const LogSet *logs);
 
 bool FindDiscGeometryFromMIMEType(DiscGeometry *geometry,
                                   const char *mime_type,
-                                  size_t file_size,
+                                  uint64_t file_size,
                                   const LogSet &logs);
 
 //////////////////////////////////////////////////////////////////////////
